feat(suite-native): Navigation, add icon, atom action icon button
