+++ packages/suite-native-atoms/src/index.ts
@@ -15,6 +15,7 @@ export * from './Modal/BottomModal';
 export * from './TipToast';
 export * from './Button/Button';
 export * from './Button/IconButton';
+export * from './Button/ActionIconButton';
 export * from './Select/Select';
 export * from './Select/SelectItem';
 export * from './VStack';
